Apply review comment and move tagChange validation from StatusChangeLogIterator to UpdateToTagCommandStep.
