+++ liquibase-standard/src/main/java/liquibase/changelog/StatusChangeLogIterator.java
@@ -1,9 +1,7 @@
 package liquibase.changelog;
 
-import liquibase.GlobalConfiguration;
 import liquibase.RuntimeEnvironment;
 import liquibase.Scope;
-import liquibase.change.Change;
 import liquibase.change.core.TagDatabaseChange;
 import liquibase.changelog.filter.ChangeSetFilter;
 import liquibase.changelog.filter.ChangeSetFilterResult;
@@ -11,7 +9,6 @@ import liquibase.changelog.filter.CountChangeSetFilter;
 import liquibase.changelog.filter.UpToTagChangeSetFilter;
 import liquibase.changelog.visitor.ChangeSetVisitor;
 import liquibase.changelog.visitor.SkippedChangeSetVisitor;
-import liquibase.exception.ChangeNotFoundException;
 import liquibase.exception.LiquibaseException;
 
 import java.util.*;
@@ -61,40 +58,12 @@ public class StatusChangeLogIterator extends ChangeLogIterator {
                         }
                     }
                 }
-                if(GlobalConfiguration.STRICT.getCurrentValue()) {
-                    TagDatabaseChange tagDatabaseChange = getTagDatabaseChange(changeSetList);
-                    boolean thereIsTagDatabaseChange =  tagDatabaseChange != null;
-                    if(!thereIsTagDatabaseChange) {
-                        throw new ChangeNotFoundException("TagDatabaseChange", databaseChangeLog.getRuntimeEnvironment().getTargetDatabase());
-                    }
-                    else {
-                        if(!isThereTagDatabaseChangeMatching(tagDatabaseChange)) {
-                            throw new LiquibaseException(String.format("Command execution tag %s does not match with changeSet tag %s", this.tag, tagDatabaseChange.getTag()));
-                        }
-                    }
-                }
             });
         } catch (Exception e) {
             throw new LiquibaseException(e);
         } finally {
             databaseChangeLog.setRuntimeEnvironment(null);
         }
-    }
-
-    private TagDatabaseChange getTagDatabaseChange(List<ChangeSet> changeSetList) {
-        TagDatabaseChange tagDatabaseChange = null;
-        for(ChangeSet changeSet : changeSetList){
-            for(Change change : changeSet.getChanges()) {
-                if(change instanceof TagDatabaseChange) {
-                    tagDatabaseChange = (TagDatabaseChange) change;
-                }
-            }
-        }
-        return tagDatabaseChange;
-    }
-
-    private boolean isThereTagDatabaseChangeMatching(TagDatabaseChange tagDatabaseChange) {
-        return (tagDatabaseChange != null && tagDatabaseChange.getTag().equals(this.tag));
     }
 
     //
